refactor(overlay): store ReviewContext and chrome-free diff view types
- Move DiffViewMode out of chrome.storage preferences so a CLI host can import the overlay
- Store and panes accept ReviewContext so local reviews do not need a GitHub PR identity
