+++ apps/extension/src/content/overlay/diffView.ts
@@ -1,0 +1,4 @@
+/** Unified vs side-by-side. Chrome-free so any host can import it. */
+export type DiffViewMode = "unified" | "split";
+
+export const DEFAULT_DIFF_VIEW_MODE: DiffViewMode = "split";
