+++ apps/extension/src/content/overlay/useOverlayKeyboard.ts
@@ -27,7 +27,7 @@
 import { useEffect, useRef, type MutableRefObject, type RefObject } from "react";
 import { getConfirmationDialogElement, isConfirmationOpen } from "@extension/lib/confirmation";
 import type { SelectableLine } from "./commentTypes";
-import type { DiffViewMode } from "@extension/lib/preferences";
+import type { DiffViewMode } from "./diffView";
 import { trapTabKey } from "./focusTrap";
 import { useReviewStore } from "./store";
